Step 1: Read and Print
A couple of architectual decisions:
- I decided to implement each Mal data type as its own data type instead
of using es2015's native data types, albeit each of these data types
uses es2015's types to represent its value.

- I modified the initally suggested regex tokenizer from
/[\s,]*(~@|[\[\]{}()'`~^@]|"(?:\\.|[^\\"])*"|;.*|[^\s\[\]{}('"`,;)]*)/
to
/[\s,]*(~@|[\[\]{}()'`~^@]|"(?:\\.|[^\\"])*"|;.*|[^\s\[\]{}('"`,;)]+)/

The only difference between both regexp is tockenizer[-3], in the original
regexp it uses `*`, but that allows empty matches('').

My modification uses `+`, which prevents any empty match.

The consequence of not making this change is that in the tockenizer function
when running `tokenizerRegexp.exec(str)` if there is any empty match(''), for
some reason this empty match never gets consumed and `tokenizerRegexp.exec(str)`
never returns null.

- At some point I considered to implement a `toString` method for each of the
MalTypes. That would have simplified the implementation of `pr_str`.
At the end I decided to make `pr_str` as explicit as posible and not "hide"
its implementation on the "internals" of each MalType.

- At some point I considered giving each of the compound data types start and
end chars a named variable, but that ended up being very verbose and the code
was less readable (I had the same experience when trying to name the reader
macro chars).

Thoughts:
The regexp tokenizer is pretty magical, it takes almost all of the work out
from parsing a string and dividing it into its own tokens.
The bad side of it is that it is completely unintelligible at a first glance,
so the only way that I could understand it was by "dividing and conquer".
I had to comment each of its sub-regexp in order to make sense of what was
going on.
It would be great if there was a way to compose a regexp from smaller named
regexp.